Optimize coercive parameter typechecking
Constraints are rather easy to typecheck against. It is sufficient to
typecheck against coercion target and constraint types. They can, in
turn, be other nominalizables with complex typechecking. But at least
one level of invocation can be eliminated. The optimization makes
loop in the following example to run 8x to 9x faster:

    sub foo(Bool() $v) { !$v }
    for ^10_000_000 { my $r = foo(False); }