Add wait overload with predicate to Lock::ConditionVariable
Previously, the wait method of Lock::ConditionVariable follows the
paradigm of the C functions. This adds a more rakuish variant that takes
a predicate. Much like the protect method in Lock, this makes it much
easier to do the correct thing (and people will do things incorrectly
if the easiest way is not also the correct way).