Add support for userspace encodings in Blob.decode
This adds support for decoding Blobs with userspace encodings
registered in Encoding::Registry.  I initially thought that we could
do this by using the Encoding::Registry decoder in all cases (both
builtin and user-defined) since it the registry should include all
builtin decodings.

However, due to what appears to be a bug in MoarVM, the version of the
builtin decoders in the registry does not correctly normalize some
encodings into NFG.  I have opened an issue regarding that bug.
see: https://github.com/MoarVM/MoarVM/issues/1443

In the meantime, I wrote this code to use the builtin decoders when
decoding a supported encoding and to try the registry only if passed
an unknown encoding.

Because this change involves decoding, I ran the full stresstest.  I
have also prepared accompanying Roast tests for inclusion in
S32-encoding/registry.t.