Remove utf32 from encodings list for MoarVM
Rakudo currently lists utf32 as one of the encoding/decoding
operations it supports.  However, this support has never been added to
the MoarVM backend and attempting to use it throws a somewhat
confusing exception.

Accordingly, this commit removes utf32 from relevant lists in
Rakudo::Internals and Encoding::Registry when Rakudo is compiled for
MoarVM (the JVM backend supports utf32).  This causes spectest
failures because Roast was checking whether an encoding was listed in
the Encoding::Registry without testing whether the encoding could
actually be used.

I have prepared a companion PR that marks these tests as todo for
rakudo.moar and that adds substests to ensure that registered
encodings can actually be used.