Make .keys/values/kv/pairs/antipairs/invert work on Enums
It appears to be a recurring WAT when people try to find out which
possibilities there are when using a specific enum.  Before this
commit, "say Bool.keys" would show "()", after this commit it will
show "(True False)" (or "(False True)" depending on hash order).

This is achieved by adding a check for enums to the Any:U candidates
of these methods, and then calling the appropriate method on the
.enums method.

cog++ for the nudge