Make sorting up to 2x as fast
Following up on a question on StackOverflow:
  https://stackoverflow.com/questions/66496614/raku-slow-mixes-sorting
it turned out that checking whether an Order enum is not Same, would
create a Bool object, which would then be converted to a native int.
Research is going on into fixing that (bypassing the Bool object), but
it occurred to me that checking for an Order::Less or Order::Same object
using nqp::eqaddr is going to be more efficient anyway.

So the sorting routines in Rakudo::Sorting were adapted to check for
specific Order enums.  This broke the Str.split(@needles) function, because
the sort block there returned a native int.  Added a sorter function that
depends on having the comparator return an Int, and adapted the generic
sorter to check for a return signature of something Int / int.

Checking for Order enums in the sort logic, forced this sorting logic to
be compiled *after* the Order logic in the setting: a re-configure is
therefore needed after this commit.  This also needed Rakudo::Sorting to
be punned in in Any-iterable.