Take a little bit of different approach on sorting comparators
Instead of checking for Inty return values of the comparator, and
Order return values, just check for Order values being returned by
the comparator, and make that take the fast nqp::eqaddr path.

Otherwise, use the old generic code path, but instead implicitely
asking for a Bool of an Order enum, directly coerce it into a
native int (which bypasses the issue that made cmp slow in the first
place).

This should now just allow Bools and other numeric values to be returned
by comparators.