Add --> Order:D to protos of <=> cmp leg unicmp coll
So that using these as comparators with sorting will take the fast
path.  Also remove any --> Order:D from other candidates of these
infixes, as they are not needed there anymore.

This caused the movement of some protos and candidates, as they
should now appear *after* the Order enum has been created in the
setting.