[WIP] Add snapshot tests for current USAGE msg
As part of my effort to refactor the USAGE message generation, I want
to ensure that all changes to existing usage messages are
intentional.  To check this, I am adding both a snapshot test that
tests new usage messages against previous usage messages and a new
tool that re-generates the snapshot.

This current version provides fairly full coverage of usage messages
when &MAIN is a single sub, but does not yet test &MAINs that are
multis.