Make signal handler setup synchronous to avoid race conditions
On MoarVM setup of a signal handler is itself an asynchronous operation. Wait
for it to complete before returning as not only our spec tests but probably
every single user expects the signal handler to be in place at that point. This
was already true for the JVM and JS backends.

Fixes a race condition in t/spec/S17-procasync/kill.t