Prepare multi dispatch for new-disp resumption
We'll use the resumption mechanism not only for `callsame` and the like,
but also as a means to try candidates with `where` clauses and the like
with only a single invoke, using a NYI mechanism to be added to MoarVM
in order to turn bind failures into a dispatch resumption. This re-works
the new-disp multi dispatch handling in preparation for both of these,
while trying to keep the cost low for the common case of a multiple
dispatch that can be decided on nominal types only and never resumes.