[JVM] Unbreak method ords for Str
For unknown reasons using native arrays in the core doesn't work.
It looks like the array is created as VMArray and not VMArray_i.
The resulting error is (in a couple of spectest files):

  VMArray representation does not implement push_native

I'd say we are still in the 'make it correct' phase for the JVM
backend, so it doesn't matter if the code runs a bit slower (by
not using NQP ops).