Make backtraces consistent on all backends
Usually backtraces are created from exceptions. To catch such an exception the
BackTrace constructer used a try thunk which at the VM level got its own call
frame. However MoarVM's specializer may decide to inline this frame. While
recent improvements to MoarVM made backtrace generation inline-aware, there
still remains a problem when the inlined frame and the nqp::backtrace op are
situated in the same frame.

As an alternative, MoarVM now supports calling nqp::backtrace with nqp::null
instead of an exception and that can be used to get consistent backtraces.
With this we finally have a definitive answer on how many frames we need to
skip to arrive at the frames that are interesting to the user.

This fixes Failure backtrace related spec tests when run on MoarVM with
MVM_SPESH_BLOCKING and MVM_SPESH_NODELAY and also when run on JVM.