Fix race condition in ThreadPoolScheduler
t/spec/MISC/bug-coverage-stress-6.d.t died occasionally with "continuationinvoke
expects an MVMContinuation" when run on a heavily loaded system. The reason for
this is that a different thread may have taken the continuation job from the
queue before the continuation was actually taken and stored in the variable.

The comment above the code even points out this risk, but the lock that got
introduced for fixing it was not enough. Ostensibly the $l.protect before
pushing the continuation job onto the queue should serialize control flow.
After all at that point the lock is still held. However Lock is reeantrant and
if protect is called on the same thread that's already holding the lock, we
do not wait but continue processing.

To fix this race condition, acquire the lock in the code block pushed onto the
queue. We can unlock immediately as the lock will not be used anymore anyway.