[JVM] Limit special cases to variable declaration
We have a couple of workarounds for native arrays in the setting not
working correctly on the JVM backend. Having something like

  my int @arr;

was compiled to a VMArray instead of a VMArray_i. It turned out that
binding to a new native array does work:

  my @arr := array[int].new;

By using this construct for the JVM backend the scope of the needed
special casing is reduced. (Also a really bad hack was removed from
src/core.c/ThreadPoolScheduler.pm6, were we relied on the mentioned
miscompilation.)