Introducing $*RAT-UPGRADE-CLASS
This dynamic variable indicates the *class* in which an UPGRADE-RAT
method will be called whenever a Rat has a denominator that does not
fit in a 64bit native integer.

The default setting for $*RAT-UPGRADE-CLASS is Num: this will silently
downgrade a Rat to a Num (floating point), thus sacrificing precision
for speed (the current behaviour).

Other possible settings are:

- CX::Warn    downgrade to Num but warn
- FatRat      silently upgrade to FatRat
- Failure     don't upgrade, return an appropriate Failure
- Exception   don't upgrade, throw an appropriate Exception

Although the dynamic variable indicates a "class" is expected, there
are no technical obstacles to put anything else in there, as long
as it can dispatch to a "UPGRADE-RAT" method that takes two Ints:
the numerator and denominator.

This does *not* affect the inlineability of the CREATE_RATIONAL_FROM_INTS
subs, but it does make the default case of silently downgrading to Num
slightly slower.  Currently this slowness is mostly caused by &DYNAMIC
being slow, and no locally cached dynamic variables (which will hopefully
change in the not-so-distant future).

Inspired by the discussions at:

  https://www.reddit.com/r/rakulang/comments/mipog4/raku34_python19_extreme_math_steve_roe/
  https://twitter.com/liztormato/status/1378294510430601220