[JVM] Support backtrace creation from nqp::null
This makes the JVM backend behave like MoarVM when creating a new
backtrace without an existing exception. For more details see
https://github.com/rakudo/rakudo/commit/2e79780e4a.

Please note that the change for src/core.c/Deprecations.pm6 fixed
an oversight in the above commit. (Now only relevant for the JS
backend, but this lead to spectest failures in S02-types/isDEPRECATED.t
on the JVM backend before.)