Remove uniprop-(int|str|bool)
They appear to have been added at some point, but they are not tested
nor are they documented, nor are they needed in the setting nor any
core modules.

Their functionality seems to intended as a performance shortcut for the
more general "uniprop" and "uniprops" functions.  My argument would be
that *if* one knows about these, one might as well use the underlying
nqp::ops as these appear to be a very thin HLL wrapper for them.