Make sure Allomorph.substr-rw returns Str
This was a little more involved.  First of all, the Cool: candidates
missed a the invocant type:D, which caused a weird interaction with
the sub versions.

Finally fixed with adding an implementation-detail candidate that takes
an additional type with which to create the string.  The other candidates
have been adapted to call that candidate with the intended type.

Also added some more candidates for better dispatchability.