Speedup exponentiation by positive Ints
Before, `my $a; $a := $_ ** 2 for 1..10_000_000; say now - INIT now;`
took ~7.1s, now it takes ~5.7s. However, `my $a; $a := $_ ** -2 for 1..1_000_000; say now - INIT now;`
took ~2.3s and now takes ~3.7s. The slowdown in the -2 case is because
of the non-default (smaller) epsilon passed to .Rat, but without it
`1_000_000 ** -2` would result in 0 instead of 0.000000000001.