Fix compilation failure with EVAL in precompiled module's mainline
The error was reported as "Cannot look up attributes in a NQPMu type object"
with a require in a module's mainline. The error message gets reported because
QASTCompilerMAST cannot find the frame for a given CUID. The error only
appeared when the require loaded an already installed and precompiled module
and only if the repository chain got changed since (by way of 'use lib' or -I).
The reason is that in this case we re-check dependencies and have to run an
EVAL to create the DependencySpecification object.

The EVAL got compiled and ordinarily the created frame would be added to
%*COMPILING<moar><mast_frames> but we cleared %*COMPILING before loading a
precompiled module. QASTCompilerMAST would therefore initialize a new
mast_frames hash to add the frames to but this hash would not be known to the
outside compiler.

The code for clearing %*COMPILING originated in the very first implementation
of the export trait in 2011. Back then we had a $*COMPILING flag and the export
trait would only do something if that flag was set. This was a hack required
because traits were also run when loading a module. The flag was later replaced
by %*COMPILING. The hack itself has been removed already in 2012 in a commit
aptly named "Eliminate old hack from the bad old days when we ran traits again
at startup". However the clearing of %*COMPILING remained and has since been
faithfully ported through all our module loading refactors.

Lay the remnants of the original hack to rest, finally.
Fixes GH #3749