Make Rakudo::Iterator.Empty about 80x as fast
The BEGIN block prevented it from being inlined.  Moved the compile
time value into a constant, and now it does inline.  And the bytecode
being very small indeed (14 bytes) it will inline quickly.

This method is called in many places where it is clear from the
beginning that an iterator will never produce any values.