Fix SlippyIterator.is-deterministic
Make it always return False.  As the class consuming the role can have
multiple slippyiterators during its lifetime, as soon as one returns
False, the whole of the iterator should be considered non-deterministic.

But that would require a time-machine, which we don't have.  So do the
safe thing, by marking an Iterator class consuming the SlippyIterator
as non-deterministic.  If that class can ensure that it always *will*
be deterministic, it can always override the "is-deterministic" method.

Fixes #4335, Brad Gilbert++ for spotting