Skip tests for NativeCall and Telemetry on JVM
The modules NativeCall and Telemetry currently don't work on the JVM
backend. I'm trying to work on this, but progress is slow and
uncertain. Open issues:

* https://github.com/rakudo/rakudo/issues/3043 (NativeCall)

* https://github.com/rakudo/rakudo/issues/4354 (Telemetry)

To avoid 'make test' spitting out a lot of failures for tests in
t/04-nativecall and t/06-telemetry, I'm disabling those dirs for the
JVM backend. The idea is to revert this, once the above issues are
solved.