Make DateTime.posix about 12x as fast
While looking at the code, and how it handled the "ignore timezone"
option, it occurred to me that `posix` actually doesn't care about
leap seconds.  And the timezone is expressed in seconds.  So the
quick way to get a posix value is just to subtract the timezone
value from the result of the Claus Tøndering algorithm.

Also make DateTime.posix a multi, so a. it can easily accomodate
other candidates, and b. to make the common case (do *not* ignore
timezone) much better optimizable.