Add a `repl` sub to enter the repl from within a program (#4325)
This adds a "repl" sub to Rakudo, that will enter the REPL at the point it is called.  Typing "exit" will then just leave the REPL and continue with the program.  One could consider this functionality like a mini-debugger, as it can inspect any variable / object within its lexical scope.