Tweak worker addition for many small tasks
Today we add additional workers to the thread pool based upon lack of
progress. This works well when work items are relatively large, and so
hold on to a worker for a while. For applications that schedule lots of
small pieces of work that are well below the threshold at which the
supervisor runs, however, some work is always completed, and so the
condition for adding further threads to the pool is not met. This could
be observed by hitting a Cro HTTP server that does relatively little
work to produce the response with a lot of concurrent requests, and
seeing nearly no workers being added, and CPU usage staying very low as
a result.

With this change, we look at the amount of outstanding work also. If
there is more work outstanding than workers available, we add an extra
worker. This has been observed to lead to additional worker creation in
a Cro HTTP server under load, which in turn improves request rates.

The intuition behind considering queue length relative to number of
workers is trying to reach enough workers to be in a steady state -
that is, we're processing work items roughly at the rate they are
arriving at. If there's consistently one next work item for each
worker, and it's not growing, then we're keeping the workers busy,
while simultaneously not having more than we need to keep up.

This is an easy approach to explain, and is cheap to implement. It may
have a tendancy to create workers a little too eagerly if there is a
one-off burst of tasks early on, so we could perhaps `min` the number
of workers with some value, however it's not immediately clear what
value to pick, and the cost of one more worker than needed is not all
that high.