Make .grep(Regex) and .first(Regex) about 40% faster
By not needing to create relative expensive .Match objects when only
a Bool can do.

For historic reasons, Regex.ACCEPTS is *not* expected to return a
Bool (contrary to all other .ACCEPTS methods) but a Match object.
This makes sense in the context of "foo" ~~ / bar /, but it does
*not* make sense in the context of a .grep or .first, where only
a Bool is needed.

Achieve this by creating a Regex.ACCEPTS-AS-BOOL implementation-detail
method that just uses the validity of the cursor as an indicator
(similar how .contains(Regex) does this).