Add .Buf / .Blob coercers to Blob / Buf (#4374)
Add .Buf / .Blob coercers to Blob / Buf

This is a bit of a special case, as Buf / Blob are both *roles*, yet
they are generally used as classes with their default parameterization.

Therefore, the .Buf and .Blob coercers return the type **with the same
parameterization** as the source.  This feels like the most logical
thing to do.

An alternative would be to add .immutable / .mutable methods to Buf / Blob.
This could then maybe also be used for turning a Hash into a Map and an
array into a List (or even, remove all containers from a List).

Note that because it is apparently impossible to actually stub a role in
the setting, the lookup for Buf is done at runtime.