RakuAST handling of regex sigspace
There were a few ways this might have been designed, but modeling it as
a node wrapping the piece of regex that is followed by sigspace seems
the least error-prone. A flag on every regex node that it's followed by
sigspace is more memory intensive in the common case, and not always
meaningful anyway, plus it would be a nuisance to apply it everywhere.
It does force all of the decision-making into whatever is constructing
the AST nodes, but so will force some whitespace insertion onto any
pretty printers too (as well as making sure they already think they are
emitting sigspace code, if they want to properly handle synthetic
constructions).