Implement default arguments in RakuAST
This is the first consumer of the expresison thunking support: since we
need to install the default value producer into the Signature object for
non-trivial cases, we need to thunk it. Much of the infrastructure for
building that thunk is factored out, so hopefully dealing with further
such situations will be relatively straightforward (some more than
others).