Introduce [1,2,3]:Int syntax
Arrays created with [] do not have typing.  It is one of Raku's traps
that people do: sub foo(Int @a) { }, and then get a typecheck error
when they pass [1,2,3] as an argument.  This commit would allow
specification of [1,2,3]:Int as a valid argument.

This implementation uses the fact that you *can* specify adverbs on
operators.  It basically looks if there is a single adverb, takes
its name and does a type-lookup for that.  And then uses that type
to create the correctly constrained Array.

Because we cannot have multi candidates that take *no* named arguments
and take any named arguments, this implementation affects the performance
of the standard logic of [].  So this should probably not be merged as
such.

One way around that now, would be to only allow for classes in Core to
be used as a type: then we could add specific candidates for these.

Another solution could be to think of completely new syntax for this
feature.