Implement Foo.pick(**)
This will produce all values of Foo.pick(*), and then produce all values
of Foo.pick(*) again, and again, ad infinitum.

This removes some unnecessary protos.  And adds a Rakudo::Iterator.Reiterate
method to produce values from a Callable that produces an iterator.

The Any.pick(HyperWhatever) candidate needs to be marked as "default"
because of ambiguity with List.pick(Any).