Introduce "done foo" as shortcut for "emit foo; done"
As discussed in https://github.com/Raku/problem-solving/issues/249

This makes the "done" sub a multi, and adds logic in the "run-done"
handlers in Rakudo::Supply to handle anything specified that is not
an exception.

Should we want to be able to emit an Exception, then we will have to
add some more logic for handing the QUIT phaser's handling of "run-done".