Fix issue with grep when storing result in array
Basically, the grep wrapper was only working correctly if the iterator
was just being called with "pull-one", but would fail to end if the
"push-all" method was called on the iterator (my @a = foo.grep...).

The fix was to have the LAST handler in the grep wrapper for Callables
throw another "last" with the correct value to be handled by the outer
handler.