Fix unimatch for non-base properties
Prior to PR #4172, unimatch worked for non-base properties such as

    unimatch '[', ']', 'Bidi_Mirroring_Glyph'; # True

That PR, however, provided a speedup for base properties, but broke
the more advanced use.  This adds back support for the more advanced
use case as a fallback _after_ testing for a base property (which
should avoid performance impacts in most cases).