Make $() work properly on scalar values
- let Mu.item return a container if there is none yet
- make sub item() actually return containers (is raw was missing)
- make the sub item version call the method item, for consistency

All spectest clean.

This now makes working with sigilless variables easier:

    my \a = $(42);
    say a;  # 42
    a = 666;
    say a;  # 666

Cherry-picked because Chemistry::Elements won't be erroring out
anymore, which was the reason for the revert of this commit in
2021: https://github.com/rakudo/rakudo/issues/4459 .  Still
spectest clean as then, and will actually allow some cleanup
in code, as some classes have their own .item with these
semantics.