Add 'modf' to the Raku core's numerical routines
This routine is commonly found in many popular
programming languages (including Python) but
was not included in the original Raku release.

The routine as added follows the basic implementation
as found in C and C++ with some slght modifications
to take advantage of existing Raku features.
As implemented in Raku routine 'modf' provides:

+ for any number input it returns a list of the integral and
  fractional parts
+ both parts retain the sign of the input number
+ if either part is zero, that part is coerced to an integer
  zero
+ the returned integral part is truly an integer (as opposed
  to some other languages such as C and C++)
+ an optional second parameter ('places') can be entered to
  force the fractionsl part to the desired number of
  decimal places (useful for testing)