Implement `last` with value (#4415)
Introduce "last foo" to end loops *and* return a value *and* FIX nodemap / deepmap / duckmap's handling of "last".

This adds the exception checking logic to all of the standard .map
functions.  Basically, three cases where to be checked:

- pull-one
  Return any non-null value from the exception and reset the iterator
  to produce an IterationEnd on the next call.

- push-all
  Add any non-null value from the exception.

- sink-all
  No change, we weren't interested in the value anyway

Note that passing a `Slip` as a `last` value, *will* produce the values separately, as one would expect.