Allow CTRL-c to stop entry in REPL (#4405)
Allow CTRL-c to stop entry in REPL, and a second CTRL-c to exit the REPL if nothing else was entered.

This in response to #4402.

This does *not* achieve the suggested behaviour of being able to
stop a runaway piece of code, as Raku's signal handler runs on a
different thread and we currently do not have a way to force an
exception on anything else but the current thread.  It *does*
allow one to restart entry without having to remove all characters.