Implement "next foo"
Allow a value to be specified with "next".  In a .map like context,
it specifies the value of the current iteration, while skipping the
rest of the iteration (as you would expect "next" to do).  In a .grep
like context, the value specified is inspected on truthiness: when
true, then the element will be included (while skipping the rest of
the iteration).  In any other case, the element will *not* be included.