Fix some errors WRT curried role typecheck lists (#4436)
Handle updates to curried role typecheck lists more lazily

The following throws upon composition of Bar:

role Foo[Any ::T]         { }
role Bar[::T] does Foo[T] { }

This happens because in order to compose Bar's role typecheck list, it
needs to know Foo's as well, but this completes the parameterization of
Foo before T has a chance to get instantiated, making it a Mu during
selection of a candidate for Foo.

Metamodel::CurriedRoleHOW.complete_parameterization and related
metamethods assume that the role typecheck list needs to be updated at
the same time. These are the only places the update_role_typecheck_list
metamethod is called, implying the role_typecheck_list getter that
depends on it needs to update this beforehand for one reason or another,
which makes sense; adding roles as late as it does should invalidate any
typecheck list cached. If completing the parameterization here is
problematic, then defer it by allowing the cache be cleared and updating
it when need be.