Narrow down test to loop phasers
When mapping / looping, the test for phasers was too large: only the
NEXT / FIRST / LAST phasers have effect on the way a loop should run.

This improves code like:

    my int $a; for ^100000 { my @a = do for ^10 { ENTER ++$a } }

by about 15% because it does not need to check for phasers at setup.

This introduces a "has-loop-phasers" method, which should probably
be marked as an implementation detail.