Improve performance of .grep(Callable) and cleanup
- about 1% for single arg Callable
- more than 2x as fast for 2 arg Callable
- a few percent for 2+ arg Callables

This gets rid of the CheatArity hack, that turned out to be only in
place to let "sequential-map" select the correct iterator to use.
Selection of the correct iterator now happens in !grep-callable itself,
so that we don't need the CheatArity hack anymore.

The 2 arg Callable's improvement is due to the fact that it no longer
needs to take a full capture, but instead operates on a Callable that
just takes 2 parameters.