First steps for Proxy handling in multi dispatch
Dispatch programs are limited to operations such as capture transforms,
guards, and attribute reads. When we have a value in a Scalar container,
and we need to extract it to determine a multiple dispatch, we can read
the value attribute and be done. Reading a proxy involves running code,
however.

If while considering a multiple dispatch we spot that some arguments are
Proxy (actually we don't look for it specifically, and so remain open to
other kinds of container in the future), we instead dispatch to a bit of
code to read the values from the proxies; it then resumes the dispatch
process, and we can use the read values to calculate the candidates that
apply. The original arguments must be passed along during the invocation
of the candidate, since there may be an `is rw` argument that wants to
receive the Proxy. (This needs more testing, and may not fully work
yet.)

The code that strips away the Proxy values will be generated eventually;
for now we handle just a small number of cases by pre-filling the table
of proxy readers (which is a good idea for the cases that CORE.setting
depends on anyway, plus this should cover operators).