Reimagine IterateTwoWithoutPhasers
This affects for loops / maps that take 2 parameters.  With regards
to performance, there's very little difference with the push-all and
sink-all state, although they have fewer allocations now, so that may
affect things in a real life situation.

The pull-one case is about 1.5% faster, also having fewer allocations:
the special part is here that the logic for handling the final iteration
when there are not enough values, is handled by a private method.  Since
the most common case is that an iterator produces exactly the number of
values needed, the common case this way has a much smaller bytecode
footprint.