Re-imagine IterateOneWithoutPhaser.pull-one
Makes that method about 10% faster.  But in most cases it won't get
called, as these simple types of for loops / maps are codegenned as
calls to the p6forstmt op.  But it's good to see the fallback being
a little faster.

In any case, the code is smaller and simpler.