Correct a test relying on a changed impl detail
Prior to new-disp, mutliple dispatch worked by doing a trial bind of the
signature of a multi candidate containing a `where` clause. If it passed
then the candidate would be invoked "for real". This means that `where`
clauess would get evaluated twice on a successful candidate.

The new-disp based multiple dispatcher works differently. It calls the
candidate in a mode where a binding failure will turn into a dispatch
resumption, which triggers trying to next candidate. However, if it is
successfully bound then it just continues running. This means we no
longer have to run the `where` blocks twice (an obvious performance
advantage!)

Having `where` blocks that do side-effects is very much a case of "on
your own head be it". This test relied on the side-effects in order to
test closure semantics, and thus needs updating.