Ensure proxy readers are marked as thunks
Somewhat unfortunately this means we have to get rid of the blocks we
prepare ahead of time because there's no way to get those marked as
thunks. More unfortunately, this breaks the reproducible builds test
(although in a sense this is good, since less trivial programs would
have had problems with build reproducibility and we'd not have known
about it so soon).

This is needed so that they get skipped when walking the call chain.