Implement p6sink using a dispatcher
The dispatcher is a no-op not only in the case that the value is not
concrete or lacks a `sink` method, but also - newly - when the sink
methood is exactly that in `Mu`, which we assume is a no-op. This should
get rid of numerous sink method calls. It's also the last place in
Rakudo itself that a legacy invoke op was emitted.