Make sure .grep knows about next/last foo
Because the .grep logic lives in 6.c, it would be calling 6.c's next/last
that don't accept any value.  Fix this by directly calling the exception
throwing logic, which *does* work in 6.c.  This might actually have some
performance benefit as well, as it removes one level of indirection.

Also normalize the order of signal code to NEXT/LAST.

This fixes #4467, melezhik++