Make Uni cmp Uni work
Also change the infix:<cmp>(@a,@b) candidate to infix:<cmp>(List:D,List:D)
so that we can be assured that only actual Lists will take that candidate.

This also makes cmp work on native arrays, albeit through the long and
slow way of making them Lists first.  To be fixed in consecutive commits.