Add Iterable cmp Iterable candidates
If either side of an infix:<cmp> is an Iterable, then use the iterator
on both sides to determine the result, rather than stringifying the
Iterable(s).  This is, I think, more correct, and much more memory
friendly as there is not Str representation created of the argument.

This change is spectest error free.  It however changes the outcome
of:
    dd "a b" cmp ("a","b")

from: Order::Same (which I consider to be incorrect), to Order::More.

The "is default" is needed to prevent ambigous dispatch errors with
the Any / Real and Real / Any candidates.