Add Uni.ACCEPTS(Uni) candidate
so that Uni(97) ~~ Uni(97) returns True.

Did not add additional candidates for non-Uni, as I'm not 100% about
the semantics, *and* it's going to be tricky to implement in the setting
due to the late definition of the Uni class, and the fact that the Uni
class has a "is repr" trait.