Refactor signature-signature typechecking for efficiency
ACCEPTS on a signature given a signature rather repetitive in that it
needs to iterate over each parameter in both signatures before
performing any typechecks, demanding a second iteration in all cases. If
a schema for the RHS is built by passing over its parameters once, then
typechecks for each parameter in the LHS can be shortcircuited sooner,
leaving any extraneous parameters from the RHS to pass over once more.